# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Oficina Backend API",
        "description": "Documentacao base da API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Listar serviços do catálogo",
                "operationId": "e9c245b5971f520d65324305d333ec2a",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de serviços paginada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Cadastrar novo serviço",
                "operationId": "945a10af0257bf35951751c077b475cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "base_price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Troca de Óleo"
                                    },
                                    "base_price": {
                                        "type": "number",
                                        "example": 80
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "estimated_minutes": {
                                        "type": "integer",
                                        "example": 30,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Serviço criado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Buscar serviço por ID",
                "operationId": "c147cd0e6c39b6fedda1264158981a27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do serviço"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Atualizar serviço",
                "operationId": "4cf079414167d106233e6e4df4bd5d44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "base_price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Troca de Óleo Sintético"
                                    },
                                    "base_price": {
                                        "type": "number",
                                        "example": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "estimated_minutes": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Serviço atualizado"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Remover serviço do catálogo",
                "operationId": "981ed4b91f0a2f96d44a65a73308456a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Serviço removido"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token gerado pelo endpoint POST /api/auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}